requirements.txt
yfinance==0.1.70
ib-insync==0.9.73
python-decouple==3.4
pandas==1.3.3
numpy==1.21.2
